Support Upstart init system

+++ dist-assets/linux/before-remove.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 set -eu
 
-systemctl stop mullvad-daemon.service
-systemctl disable mullvad-daemon.service
+if which systemctl &> /dev/null; then
+    systemctl stop mullvad-daemon.service
+    systemctl disable mullvad-daemon.service
+elif /sbin/init --version | grep upstart &> /dev/null; then
+    stop mullvad-daemon
+fi
